requirements.txt
aiogram==3.22.0
aiohttp==3.12.15
gspread==6.2.1
google-auth==2.41.1
google-auth-oauthlib==1.2.2
